chore: move border to another file
